getbyid bugfix

+++ Models/Burger.cs
@@ -2,7 +2,7 @@ namespace csharp_burgershack.Models
 {
   public class Burger
   {
-    public Burger(string name, string description, int price)
+    public Burger(string name, string description, double price)
     {
       this.Name = name;
       this.Description = description;
@@ -14,7 +14,7 @@ namespace csharp_burgershack.Models
     }
     public string Name { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public double Price { get; set; }
     public int Id { get; set; }
   }
 }
